Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

My big gripe about this trajectory is that we are letting LinearLayoutManager places items at particular X/Y coords, and then we are allowing for modification of their position after the fact on via PageTransformers.

It seems like a much cleaner and direct path exists where ViewPager2 wouldn't use LinearLayoutManager, but would instead implement a new LayoutManager that allows for some kind of smaller component that positions views... you know?  I know that is a lot more work... but I'm a big concerned about the whole, LinearLayout will put it at (x,y) and then this other thing outside of LinearLayoutManager will put it somewhere else... and may scale and rotate it?

In fact... how are developers supposed to reposition Views in PageTransformer?  Aren't they going to fall into a weird feedback loop where our input is affected by our output?  Say the goal is to move views 50% faster.  So when the view would be drawn at 10 pixels, the dev would draw it at 15 and when it would be drawn at 20, the dev would draw it at 30.  But, I think what will happen is that it should be drawn at 10 pixels, we will instead draw it at 15, and then when it is moved 10 more pixels by the drag, it will be at 25 and we will move it to 42 (25 * (25/2)).  Right?

The old PageTransformer would only move the drawing position of the View and not the View "itself"... like the old View animation system.

-1 for now until we discuss.  Maybe I'm missing something big and am way off base.

Patch-set: 1
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1