Introduce OnTextViewStatusChangeListener
Introduces OnTextViewStatusChangeListener for compat TextViews.
The idea is that this allows "friends" of the appcompat widgets
to get and react to internal events happening inside of TextView.
With this we will be able to decouple logic for code like
FloatingToolbar from TextView.

This uses the register/unregister (add/remove) pattern for registering
listeners as setListener is insufficient for such tasks.

Test: :P
Change-Id: I8e94f8ec2aec88f4530a6da662d4a27330d33576