Update patch set 3
Patch Set 3:

we prefer not to change the default behavior of current apps, when user selects the preference to change the value, focusing on first or current selection is debatable which one is better. When it in the "back" that recreates the preference fragment, i.e. savedInstance!=null, we need make sure the focus is on current user selected value: I think that's already handled by VerticalGridView.

Patch-set: 3
Reviewer: Gerrit User 1020538 <1020538@85c56323-6fa9-3386-8a01-6480fb634889>