Update patch set 13
Patch Set 13:

I'm going to make an attempt to see if it is possible to move the FloatingToolbar to textclassifier library for now given that SmartLinkify is the only client of the toolbar atm. We can always move it back to appcompat when we have other clients for it. Wdyt?
Thanks.

Patch-set: 13
Reviewer: Gerrit User 1138403 <1138403@85c56323-6fa9-3386-8a01-6480fb634889>