MediaRouteCastDialog: Refactor UI update logic
Currently, dialog calls refreshRoutes() when onRouteChanged() is
triggered which will re-classify routes and re-positioning them in
RecyclerAdapter#setItems.

However, according to UI of MediaRouteCastDialog, route has to stick to
its initial position except selected route is changed. To fulfill this
requirement, this CL refactors UI update logic of the dialog.

- Replace update() with updateMetadataViews(), updateRoutesView().
- Since refreshRoutes() doesn't have to be called frequently like
  ChooserDialog, just call updateRoutes() directly.
- Instead, because updateRoutesView() has to be called frequently,
  add refreshRoutesView() to cumulate changes and updateRoutesView()
  not too frequently.

Bug: 113373498
Test: Manual test with support-v7-demos
Change-Id: Iebffd8b06fa3b944755aa3b08f75cd6696fdce9c