Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I believe if you do it, it is better to do it on the View level since those functions are all view functions. therefore checkbox will need it, button will need it etc etc.
> 
> If you are going to keep it private, and it will only be used by internal implementation, I don't have a strong opinion about it, but I would still prefer it not to be named "OnTextViewStatusChangeListener". If you do not add this to other classes now, I bet there will be a time where half of the widgets will have it and other half will not.
> 
> If you intend to make it public, I did not think too much about it, but I dont think it is a good idea to make it public. (I am saying this based on your comment which said "@hide for now".)

Thanks for taking the time to review and give feedback.
The idea is that the status changes will be specific to TextView. e.g. selection changed, suggestion popup shown, etc. Things that other Views won't have. Do you suggest a hierarchy of listener interfaces? e.g. View listeners in OnViewStatusChangeListener and TextView specific ones in OnTextViewStatusChangeListener? (names for illustration only).
This is not urgent or blocking atm. But if the idea seems okay to you I'll be in touch when it becomes necessary.

Patch-set: 1
Reviewer: Gerrit User 1138403 <1138403@85c56323-6fa9-3386-8a01-6480fb634889>