Adds Preference SummaryFormatter interface and default implementations
This allows preferences to dynamically configure their summary based on
the state of the preference / external factors. For example, updating the
summary of an EditTextPreference to match the text that has been persisted.

Also adds default implementations to EditTextPreference, ListPreference,
and DropDownPreference by extension. These can be enabled with
useDefaultSummaryFormatter(true) or via XML with
app:useDefaultSummaryFormatter="true".

Bug: b/113511984
Test: DynamicSummaryTest
Change-Id: Id02fd4aa549bde07673f9d2ab7930f40112ec0bf