Fix font not being applied to alert dialog title
The bug causes the title for an alert dialog to not use the font specified in the fontFamily attribute. The text for the alert dialog was of type DialogTitle which inherited from TextView and thus was not wrapped in an AppCompatTextView (which is done by AppCompatViewInflater). The class now actually extends AppCompatTextView which has support for the desired behavior.

Bug: 70818342
Test: ./gradlew appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.app.AlertDialogTest\#testTitleThemeFontIsAppliedToTitle

Change-Id: Id2bf099d6a3e8c99641df30ada49e414e2779a3f