Update patch set 3
Patch Set 3:

I'm sure you meant to say ignore less than {refVersion.major}.0.0 not {refVersion.major}.{refVersion.minor}.0  I agree that the check could be made after the result is returned, I would have done that if re-parsing the version out of the resulting file was not even more wordy and even less clear (requireMajorRevision is pretty self explanatory in a argument and the check is a matter of adding an extra check in the if.) Trying to implement the change, I realized it's a longer and more painful check, so it's not worth it.

For printing which project it is, that was the plan before writing that, but Gradle already prints the project's name while running the task so I thought it is redundant.

As for the version change, I did that, but I don't think that solved the existing problem that parseOrNull does not work properly for files and only gets the version number without the extra suffix. Maybe we can get rid of it in a future CL if we don't actually need it or use it anywhere or alternatively fix it then.

Patch-set: 3