VV2: Take over old view only when media is playing
Currently, when view type is updated, the old view is always taken
over. This creates a problem when the media is paused since the
new view shows the default image (for SurfaceView it is a black
screen, and for TextureView it is the last rendered screen). To
prevent this, this CL seeks the media to the current position after
the new view takes over.

Bug: 113360020
Test: manually & ./gradlew :media-widget:connectedCheck
Change-Id: I1ebb08463cff28bee8537a9c9e481564e62ced09