Return ListenableFuture from WorkerWrapper.
This is part of the work to allow true non-blocking work in
WorkManager.  This change encapsulates the "pre-work" part of
WorkerWrapper in a ListenableFuture that returns a single
boolean (whether to retry).  This removes the dependency on
ExecutionListeners in WorkerWrapper.

This is a messy change, with many parts of the code still in
flux.  It should be noted that this is NOT meant to be a
meaningful ending point on its own.  For example, I'm not
very happy about the ListenableFuture generic we are using.
Also, I didn't fully get rid of ExecutionListener yet.
Consider this a step on the path to a better codebase.

Test: Updated tests and ran all tests and various integration
      scenarios (Sherlock Holmes, exploding work) on APIs 16,
      19, 23, and 28.

Change-Id: I6a3aa50aee3207d72477f22073556878dfe116a4