Updates to TextLinks.addLinks(...) APIs.
Removed addLinks(TextView, TextLinksParams).
Clients should use the version that takes a Spannable instead.

Reason: When the TextView is saved and recreated, the links are
reapplied but the LinkMovementMethod is not set on the TextView.
This causes the links to appear in the TextView but not clickable.
The goal is to encourage developers to be mindful of when they set
a LinkMovementMethod on their TextView. e.g. When it is inflated.

Also changed addLinks(Context, Spannable, TextLinksParams) to
addLinks(Spannable, TextClassifier, TextLinksParams).
Clients should be mindful about the TextClassifier they decide to
use to classify their clicked links. Typically the one they used
to generate the links in the first place.

Test: ./gradlew textclassifier:connectedCheck --info --daemon
Change-Id: I816840a6c83a7c6e645d5b6fb2fa8bbeae3309f7