Support for testing different dependency versions
AndroidX libraries should now declare their dependencies as a module
with a fixed version instead of a local project (e.g
"androidx.collection:collection:1.0.0" instead of ":collection".) The
version specified should be the minimum supported version the library
depends on.

We approximate the testing of this compatibility by creating a task
that builds the library against the earliest version of its
dependencies, and another task that builds against the version in the
current development project.

Test: manual testing.
Test: testing to ensure that we only substitute fixed version with
current projects that actually exist and are within the same major
release.

Bug: 112531649
Change-Id: Ic3539e2bbf552ebcc4559c34f8724e9659c968f1