The core support library now has a third build variant which builds against latest dependency versions.
We want androidx libraries to be able to declare a minimum support
version of each of their androidx dependencies, and then be able to test
against both the minimum required version and current last release
version within the supported version range (which would generally be the
minimum required version up to the last minor version in the same major
release.) We add two build flavours called useLatestDependencyVersions
and useMinimumDependencyVersions which would build against the last
dependency versions by not enforcing using the minimum required version
since gradle will automatically use the last released version in the
specified range, or build against the minimum dependency versions.
We specify the ranges in the dependencies in build.gradle using brackets
so the dependencies will look something like this
dependency_group:dependency_name:[x.y.z, a,b,c). This means we have to
modify the androidx library build files to both specify version ranges
for their dependencies as well as add the two build flavours.

Test: manual testing with ranges as well as setting minimum dependencies
to non existant libraries both in the build dependencies and in code
with the support library plugin and getting the expected errors to
prove that we do indeed consider only the minimum version in non
useLatestDependencyVersions variants.

Bug: 112531649
Change-Id: Ic3539e2bbf552ebcc4559c34f8724e9659c968f1