The androidx libraries are now able to build against both minimum and maximum dependency version.
We want androidx libraries to be able to declare a minimum required
version of each of their androidx dependencies, and then be able to test
against both the minimum required version and current last release
version within the supported version range (we are approximating the
ability to validate that the version ranges that a library claims to
be compatible with are in fact compatible.)
Libraries can specify standard Gradle version ranges (of the form
dependency_group:dependency_name:[x.y.z, a.b.c) ) for their
dependencies.

Test: manual testing with ranges \nTest setting minimum dependencies
to non existant libraries both in the build dependencies and in code
with the support library plugin and getting the expected errors.

Bug: 112531649
Change-Id: Ic3539e2bbf552ebcc4559c34f8724e9659c968f1