Support for testing dependency version ranges
Any androidx library can now declare to be compatible with a range
of versions of another androidx library via standard Gradle version
ranges. These ranges may look like:

dependency_group:dependency_name:[x.y.z, a.b.c)

We approximate the testing of this compatibility by creating a task
that builds the library against the earliest version of its
dependencies, and another task that builds the library against the
latest released version of its dependencies.

Test: manual testing with ranges
Test: setting minimum dependencies to non existant libraries both
in the build dependencies and in code with the support library
plugin and getting the expected errors.

Bug: 112531649
Change-Id: Ic3539e2bbf552ebcc4559c34f8724e9659c968f1