Workaround TextView.getTextclassifier in O platform
When TextView.getTextClassifier is called for the first time,
the cache is set to be he one returned by
TextClassificationManager.getTextClassifier(). This is wrong because
later {TextClassificationManager.setTextClassifier(TextClassifier)} will
then have no effect to this TextView.

This bug is fixed in P, we backport the fix back to O by overriding
TextView.setTextClassifier and TextView.getClassifier().

Test: Added tests to AppCompatEditTextTest and AppCompatTextViewTest
BUG: 72307474
Change-Id: Idda2189eeb3eedbf1763806903192a25d4bf8ed8