Remove Dagger from Jetifier rules
This erroneously causes newer versions to partially regress to using 2.16 which throws wacky exceptions since all versions must match. But also there are tens or hundreds of annotation processors in the ecosystem with varying levels of compatibility and it doesn't make sense for jetifier to try to track their versions to remedy it.

Test: none
Bug: https://github.com/google/dagger/issues/1245
Change-Id: Ie4a22f2ec5d9ae43f68799fc1ba06b9dc89c36ba