Update patch set 1
Patch Set 1: Code-Review+2

As per offline conversation, this is being done on a background thread, so throwing an exception at most messes up Workmanager. My request to bounce to the UI thread just to throw the exception so that crash reporting tools pick it up is something to consider for the future.

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2