Fixing the highlighted area of PreferenceCategory is not balance
If a PreferenceCategory is a searched item, the top of highlighted
area is larger than its bottom.
If the PreferenceCategory is with icon, the icon position is aligned
with top.
Reducing the highlighted area of its top and making the icon
vertically aligned.
Regarding symptom and the visual after fixing, please refer #13 of
b/110737174

Change-Id: I623fd040c5284ab4af29d079942d7c7dc5bdfb29
Fixes: 110737174
Test: manual