Update patch set 3
Patch Set 3:

It should not be jetifier's responsibility to update third-party annotations processors. Aside from the fact that it will always be doing an incomplete job with regard to the ecosystem of processors, only updating the processor artifact is incorrect behavior. All of the artifacts need to be versioned the same. The fact that using the 2.16 android processor with all other 2.15 artifacts working was luck.

Let each annotation processor deal with incompatibility and informing users they need to upgrade.

Patch-set: 3
Reviewer: Gerrit User 1232209 <1232209@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0