Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> It should not be jetifier's responsibility to update third-party annotations processors. Aside from the fact that it will always be doing an incomplete job with regard to the ecosystem of processors, only updating the processor artifact is incorrect behavior. All of the artifacts need to be versioned the same. The fact that using the 2.16 android processor with all other 2.15 artifacts working was luck.
> 
> Let each annotation processor deal with incompatibility and informing users they need to upgrade.

It's not that straightforward. If we don't bump up the version the migration will end up failing in very non-intuitive way for the developer. This is 100% failure for 2.15. I asked Hung if we can bump up only when we see version below 2.16.

Patch-set: 3
Reviewer: Gerrit User 1238792 <1238792@85c56323-6fa9-3386-8a01-6480fb634889>