Update patch set 7
Patch Set 7:

Thanks for tracking this down. It should be fixed by
https://android-review.googlesource.com/c/platform/tools/metalava/+/760099
(running through treehugger now).

Basically in non-compat mode metalava also includes methods that are implicitly included; e.g. if your public class implements an interface which specifies an abstract method, and then you inherited a concrete implementation from a *hidden* super class, this method should not be treated as abstract in your class. (There are some other similar scenarios, which is for example how StringBuilder.setLength ends up in the signature files, though it's not there in compat/doclava mode.)  There was a bug in this code; it treated some methods as coming from interfaces (methods in interfaces do not need to specify a modifier, but they're implicitly public). I ran the code manually again on paging:common and those two methods are now correctly handled (there were no other diffs in that module).

(Fixed in version 1.1.2.)

Patch-set: 7