Update patch set 1
Patch Set 1:

There's a subset of this CL that *is* compatible. Can you extract that out and we can get those changes in?

I think a portion of the generic changes are compatible, though I am not sure if Doclava understands that. We'll want to take them one-by-one and force the API change through if we need to.

Check https://wiki.eclipse.org/Evolving_Java-based_APIs_2 "Turning non-generic types and methods into generic ones" for a long description of what types of changes are safe and split those out to a separate CL.

Patch-set: 1
Reviewer: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>