Make sure adding accessibility actions to recyclerView Items works
This is the first step to deprecating all the recyclerview custom API and also making recycler view accessibility good.

Make sure we remove all actions from itemViews when they are recycled.
That way the developer doesn't have to clean it up themselves, they just have to add them in onBind.
There isn't much of a point of keeping actions around across recycles, cause they take no view parameter, so the view has to be captured in the lambda.
So someone has to recycle it.

This is a common use case for recyclerViewDelegates, I added tests to demonstrate and make sure this is functional and easy.

Test: See the unit tests.
Bug: 31493845
Change-Id: I13446b02d52653600c02bc6af19294840e281b03