Accessibility actions in viewCompat
The developer can now add an accessibilityAction using the support
library, by overriding perfom in AccessibilityActionCompat and passing
that into addAccessibilityAction in viewCOmpat.
A few parts to it,
1) Have accessibilityActions be associated with views using setTag.
2) On creation of the node info combine the ones on the view and the
delegate there.
2) On perform action, if the normal delegate call fails, try to perfom
an action using the ones tagged on the views.
3) AddAccessibilityAction takes a lambda paramter
(accessibilityViewCOmmand), which takes a CommandArguments parameter
(as opposed to a bundle), allows for stronger type checking and it makes
it cleaner
4) ReplaceAccessibilityAction has strong type checking, on the lambda
passed in.. That is done by making all fo the StandardActions a StandardAccessibilityActionCompat
which is a subclass of AccessibilityACtionCOmpat. This is public api
change, but it doesn't seem like it will break anything.

Change-Id: I705742619c2ce0343ad97965efe4c3b393b9cc88
Fix:31493845
Test: Unit tests.