Make sure adding accessibility actions to recyclerView Items works
This is the first step to deprecating all the recyclerview custom API and also making recycler view accessibility good.

This is a common use case for recyclerViewDelegates, I added tests to demonstrate and make sure this is functional and easy.

Second part to this is make sure people can remove standard actions. Store a list on viewTags, remove when initializing accessibilityNodeInfo

The recyclerView case did drive the actions API to take a View parameter for the command, otherwise we have to clear all commands on reuse, but then the expected behavior gets iffy if the developer wants to add something on onCreate or some other place....

Test: See the unit tests.
Bug: 31493845
Change-Id: I13446b02d52653600c02bc6af19294840e281b03