Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> There's a subset of this CL that *is* compatible. Can you extract that out and we can get those changes in?
> 
> I think a portion of the generic changes are compatible, though I am not sure if Doclava understands that. We'll want to take them one-by-one and force the API change through if we need to.
> 
> Check https://wiki.eclipse.org/Evolving_Java-based_APIs_2 "Turning non-generic types and methods into generic ones" for a long description of what types of changes are safe and split those out to a separate CL.

So I changed the configuration to use Metalava instead of Doclava, and now it's happy with the API changes. However, Filip's comment is still relevant and that *will* break developers. So I'm unsure if Metalava is correct in this case. Also unsure how we can avoid breaking developers without adding a separate method that uses types and then deprecating the current one - which is sad because we just deprecated the older onSetInitialValue for the newer cleaner one, which we then might have to deprecate again for an even newer one.

Patch-set: 3
Reviewer: Gerrit User 1343300 <1343300@85c56323-6fa9-3386-8a01-6480fb634889>