Update patch set 1
Patch Set 1:

i'm a bit uncomfortable to open this up but since we have the Async version open, this seems to make sense.

We should also document that if this method is called while a transaction is active, it will not do anything .... (seems bad but that is the current case as far as i remember).

@danny, any alternative recommendation?

Patch-set: 1
CC: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>