Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i'm a bit uncomfortable to open this up but since we have the Async version open, this seems to make sense.
> 
> We should also document that if this method is called while a transaction is active, it will not do anything .... (seems bad but that is the current case as far as i remember).
> 
> @danny, any alternative recommendation?

I, like Yigit, also feel a little uncomfortable opening this method up, specially since it can trigger lots of observers that might be invalidated. Maybe an alternative can be to create a public sync method that takes an observer and invalidates it if it needs to. Basically narrowing down the scope of observers that will execute synchronously. After all your use case seems to be to make sure X queried data is up to date and not to inform other queries you performed some write that they should be aware of, which is the O.G. intended purpose of the async method.

Patch-set: 1