Fix resources rewrite bug in Jetifier.
Previous CL for annotations processing (aosp/765522) relaxed the
selector for tokens in XML files. However the original filter was
preventing symbols like "?android:attr/textAppearanceMedium" to be
picked up for processing. After such token got into processing it
uncovered that we did not assert properly that we should not handle
cases with mixed "/" and "." as separators. For instance the factory
of JavaType from dot notation was supposed to assert case like this
for long time already as it is non-reversible operation.

Test: Added the xml snippet that uncovered the breakage
Change-Id: I224e274293f0c9a648892dfbb9aa6b98bb6d0881