Update patch set 2
Patch Set 2:

Hmm, the intent for this change was to make it easy to not forget to track the API but I didn't realize that it was intentional to not track the APIs before the first beta.

Now that I test updating some versions to betas and running checkApi I do find that that fails (without this change), which I guess is the important part.

So nevermind then; thanks everybody.

Patch-set: 2
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>