Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i'm not worried about which observer gets notified though. I'm just worried that it is possible to trigger it sycnhronously that might create har dto resolve conflicts (if someone got their threading wrong)

Do you mean failing if on main thread? Or more that this offers the ability to setup observers to signal on multiple threads?

I do like the idea of scoping the sync method to not trigger other observers (which may assume they're only notified async, or on a specific thread), and in practice I thinks this helps the 2nd part - observers will opt-into the weird sync behavior to run on an app determined thread.

Patch-set: 1
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>