Allow apps to specify how to present the suggested actions after the ... link is tapped
By calling TextLinksParam.setSpanFactory, apps can create their own
spans when using TextClassifier.generateLinks(...).apply(...).

When users taps on a link, textclassifier is used to classifyh the
text in the span and invoke onTextClassificationResult once it is done.
We provide the default implementation of showing the suggested actions
(i.e. floating toolbar), but apps developers can override
the onTextClassificationResult function to change the beahavior.

If the app want to further customize it, they can provide their own
implementation of AbstractTextLinkSpan.

BUG: 72307474
Test: ./gradlew textclassifier:connectedCheck --info --daemon

Change-Id: Idee970fb402a6a8d97444c92dd99f17588f33fab