Allow apps to specify how to present the suggested actions after the ... link is tapped
By calling TextLinksParam.setSpanFactory, apps can create their own
spans when using TextClassifier.generateLinks(...).apply(...).

When users taps on a link, textclassifier is used to classifyhthe
text in the span and onTextClassificationResult is invoked once the
result is ready.
We provide the default implementation of showing the suggested actions
(i.e. floating toolbar), but apps developers can provide their own
TextLinkSpan and overrides onTextClassificationResult to change the
behavior.

If the app want to further customize it, they can provide their own
implementation of AbstractTextLinkSpan and implements the onClick
function.

BUG: 72307474
Test: ./gradlew textclassifier:connectedCheck --info --daemon

Change-Id: Idee970fb402a6a8d97444c92dd99f17588f33fab