Allow apps to specify how to present the suggested actions after the ... link is tapped
By calling TextLinksParam.setSpanFactory, apps can create their own
spans when using TextClassifier.generateLinks(...).apply(...).

If apps want to override the behaviour when a link is clicked,
they can extend TextLinksSpan and override onTextClassificationResult.
For example, some apps want to start the browser immediately if it
is classified as a URL.

BUG: 72307474
Test: ./gradlew textclassifier:connectedCheck --info --daemon

Change-Id: Idee970fb402a6a8d97444c92dd99f17588f33fab