Ignore overview.html files in source jar, suppress EP checks.
The HTML files are present in revision 6 of API 28's stub src jar. And because revision 6 uses metalava to create its stub jar, inherited methods are redeclared on subtypes. This causes error-prone to think that WindowInsets defines an equals() and thus must be compared using that method rather than reference equality.

Test: gw bOS
Change-Id: Ia617f8f95cb00bb0f22922b4f539430b6d3d22d9