Ignore overview.html files in source jar, fix equality check.
The HTML files are present in revision 6 of API 28's stub src jar.

Revision 5 of API 28 added an equals and hashCode method to WindowInsets which compares equality based on value. As a result, error-prone now enforces that we use equals() instead of reference equality in the case someone returns a new instance which represents the same value.

Test: gw bOS
Change-Id: Ia617f8f95cb00bb0f22922b4f539430b6d3d22d9