Align supplemental actions to the top of the ListItem.
When a ListItem becomes very tall, there is a change that the
supplemental icon/action disappears below the fold. As a result, UX has
dictated the icon/actions be aligned closer to the top of the view in
these cases.

For ActionListItem and TextListItem, this occurs if there's both body
and title text. For SeekbarListItem, this happens if there's text.

Fixes: 116122490
Test: Verify on SupportCarDemos' TextListItemActivity and
SeekbarListItemActivity.

Change-Id: Icf3e6b2ed8fa5043f31207c20427765f4add349c