Fix font not applied to spinner widget
The bug causes the text for the title on the spinner widget popup (when using dialog mode) to not use the font specified in the fontFamily attribute for older API levels. As the Spinner widget wasn't fully backported into AppCompat, the remaining part was backported. More specifically it now uses a new DialogPopup class (instead of the framework one) which in turn uses an AppCompat AlertDialog that uses a DialogTitle (that extends AppCompatTextView) that has support for custom fonts.

Bug: 111909701
Test: ./gradlew appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatSpinnerTest\#testHasAppCompatDialogMode
Change-Id: I439c7351b5f6b705ca3aa5deb7c23cd2928f88b9