Move Fragment's ViewModelStore to FragmentManagerImpl
Instead of each Fragment being responsible for
maintaining its own ViewModelStore and having
FragmentManagerImpl have additional logic in
setting and getting the ViewModelStore, centralize
all of the management of the ViewModelStores in
FragmentManagerImpl.

As Fragment already required the Fragment to be
attached before accessing the ViewModelStore, this
move does not change any existing behavior.

Test: existing tests pass
BUG: 116324855
Change-Id: Id8b3d53c2d942f13edc0b8ee9ac65847046701c7