CLickableSpan workaround for TalkBack
To give app developers a reason to adopt the new support library API for clickable spans, TalkBack needs to consume it.
But androidx pushes to google3 have been blocked for a couple months and talkback is converging.
If we miss this release, it will be serveral months before Talkback can consume this API, so the enableClickableSpans API will be useless.

TalkBack is parsing the extras bundle themselves and unwrapping the clickable spans without any support library help.
But to perform onCLick, we initially built this code to use a resourceId, and talkback can't consume a resourceId for a framework they haven't adopted yet.
So put a temporary id in to check for the talkback case.
This is a gross hack, and should be removed ASAP.....

Test: Unit test, and before it ships a sample app with talkback.
Change-Id: If4e15f264f6da7ec5a5bd29071da5dd266fb7cd9