Add two-argument remove to map-like collections
This was added to Map in API 24 which users of ArrayMap are able to use, but there's no reason ArrayMap, SimpleArrayMap, or the primitive-specialized collection users shouldn't have access on all API levels.

Hide KTX functions which are obsoleted by members. Their disappearance from current.txt is a Metalava bug (b/116874788).

Bug: 116778381
Test: ./gradlew :collection:build
Change-Id: I083faad519db086cec9e91b685663b7de0a2ed08