Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > BanUnlistedSDKVersionComparison bans comprison of SDK_INT
> to anything other than constants in VERSION_CODES.
> 
> Woah! I would prefer the exact opposite of this rule.
> 
> The letters and names of releases are marketing gimmicks that I drink to forget. We use integers when showing what API level a method was added in on DAC. Lint shows integers when you use an API added after your minSdk version.

@alanv what do you think about this? I don't have a strong opinion.

As for lint infrastructure going to be handling a lot this, it seems unlikely that it's going to do all the checks we would like to do although it is indeed unfortunate that error-prone runs only on Java. That being said, while we continue to wait for a response from @jameslau, I wanted to put some performance numbers, a clean runErrorProne task run on my machine (using cleanBuild.sh) with custom error prone rules caused a negligible performance hit (runErrorProne with custom rules took 3:34 while it took 3:30 without them.) So even if it does not handle Kotlin, there is no immediate downside to adding them.

Patch-set: 1
Reviewer: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>