Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> can you add a test that fails w/o the fix and continues properly after the fix.
> Hopefully, we should ensure that the view is recovered.
> 
> Also, the fact that RV's child being removed from it and randomly re-added is not something recyclerfview can account for.
> layout managers keep children in certain order and RV also has an abstraction over it for animations (Child helper). if something else changes RV's children directly, it will not work and will eventaully crash.
> 
> we can talk offline, maybe there is something I'm missing.

I will definitely cover this change with some tests once you approve the general approach of this change in RV.

I believe you are not understanding it totally correct. It will use a view only when if safe to fo so. Which means view is not attached to the parent anymore in the final state. And after the end of animation it will not attach it to RV back, it had no parent before it will not have after the animation. So it happens only when Transitions are asked to animate removal of a view from RV. The view is not attached to RV anymore, but it is somewhere in cached view holders for later use. And if later RV will try to take it from cache and add it back to RV it can be impossible because the view is currently added to overlay for animation.

Patch-set: 2
Reviewer: Gerrit User 1361834 <1361834@85c56323-6fa9-3386-8a01-6480fb634889>