Update patch set 1
Patch Set 1:

This breaks existing users of the API, with no compatibility for old users. It also opens up large amounts of implementation detail I don't think we want to be public.

What do you want to change, and what specific usecase is it for? A bug on the public tracker is generally the best place to start for an API change.

Patch-set: 1
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>