Update patch set 1
Patch Set 1:

(2 comments)

I like Jake's approach for reducing allocations but we should consider how error-prone it is for a developer to have a ViewModel type that implements CoroutineScope available in their Activity or Fragment code. It seems like AsyncTask redux in terms of accidental bug potential if people start doing arbitrary `myViewModel.launch {}` in their activity or fragment methods.

Combining the close/cancel tags in ViewModel with something like:

    class MyViewModel : ViewModel() {
        private var viewModelScope = createCoroutineScope()
    }

where `createCoroutineScope()` is defined something like:

    fun ViewModel.createCoroutineScope(): CoroutineScope

and ties into the cleared signal to cancel the context seems like the way to go.

As for whether SupervisorJob is appropriate here or not, that depends on how the developer wants to handle errors. Perhaps a second create method or customization options would be appropriate.

Patch-set: 1
Reviewer: Gerrit User 1014545 <1014545@85c56323-6fa9-3386-8a01-6480fb634889>