Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > BanUnlistedSDKVersionComparison bans comprison of SDK_INT
> to anything other than constants in VERSION_CODES.
> 
> Woah! I would prefer the exact opposite of this rule.
> 
> The letters and names of releases are marketing gimmicks that I drink to forget. We use integers when showing what API level a method was added in on DAC. Lint shows integers when you use an API added after your minSdk version.

They're useful constants for avoiding potential issues with raw ints during pre-release development. You can follow the discussion on the bug, but the conclusion that we reached is that this was the simplest way to avoid bad SDK_INT checks. If you have an alternative suggestion, happy to hear it.

Patch-set: 1