Coordinate Transition's layout suppression with RecyclerView
RecyclerView already has an api to freeze layout and scroll but it doesn't expect someone can suppressLayout in another way(via hidden method ViewGroup.suppressLayout() which Transitions are using). It is expected by Transitions that when ChangeBounds is in progress no layout changing operations will be allowed, and RecyclerView's scroll is also one of such operations, so if scroll happens while some of the children is animating by ChangeBounds this scroll offset will be applied for all the elements but animated by ChangeBounds.
To prevent it RecyclerView should be aware no scroll is allowed now. I found it is a working solution to just add suppressLayout method in RecyclerView and forward the value to our logic setLayoutFrozen(). So we actually override hidden method in this way. THe only limitation I see is we have to add proguard rule to not rename this method.

Test: also added a new test for suppressLayout to be called for ChangeBounds
Bug: 33609996
Change-Id: Ic60256982e5b81434ce5af473de724e60ddd0b40