Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I don't think we should support custom error-prone rules because they only work on Java code. I believe effort is underway to port a large subset of the error-prone checks to Lint's infrastructure so they can run on Java or Kotlin code.
> 
> Our request for an update on Lint (you're +cc'ed) has gone unanswered for over a week, so I'm assuming it's not actually happening.

Does it matter? Why would we choose to invest in a tool that supports only one of the two languages we develop in when there is a viable alternative already in place? The fact they are porting the checks (and they are) was a signal, not a prerequisite.

Patch-set: 1
Reviewer: Gerrit User 1232209 <1232209@85c56323-6fa9-3386-8a01-6480fb634889>