Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I don't think we should support custom error-prone rules because they only work on Java code. I believe effort is underway to port a large subset of the error-prone checks to Lint's infrastructure so they can run on Java or Kotlin code.
> > 
> > Our request for an update on Lint (you're +cc'ed) has gone unanswered for over a week, so I'm assuming it's not actually happening.
> 
> Does it matter? Why would we choose to invest in a tool that supports only one of the two languages we develop in when there is a viable alternative already in place? The fact they are porting the checks (and they are) was a signal, not a prerequisite.

If it's a viable alternative then someone needs to respond to that email. Right now, it looks like a promise rather than a tool.

We need these checks in place. Once there's a real, supported alternative in place then we can write the checks again against whatever API we're given.

Patch-set: 1