Update patch set 2
Patch Set 2:

Also, a note for reviewers, it seems that this (exacerbates a bug and) adds +75% (yes, almost double) to the total runtime of `./gradlew buildOnServer`.

Alan mentioned verbally that that was acceptable; it still makes me uncomfortable.

I figured out how to cause it to only increase the runtime by about 1m30s by putting a fix into the Android Gradle Plugin, but they told me to submit that to studio-master-dev instead, which means if we submit this it will make the build extra-slow until we upgrade to AGP 3.3-Canary

Patch-set: 2
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>