Automatically generate api diffs
We are now able to generate api diffs for a certain library.
Added a python script that does that step for you and zips the
file so it's ready to upload in a bug and be done with diffs
step.
Ideally the script will also create the bug for you and attach
the file, however there does not seem to be support for attaching
files to a bug from the commandline (the bugged tool), once I find
a way then I'll update the script to perform that so generating
and submitting api diffs will truly become a one step process.
This is strongly related to aosp/786836 as it will let us
create diffs against previous alpha (and possibly beta) release.

Test: run the generateLocalDiffs task on a library and see
the diffs being created. Run the script passing in the appropriate
arguments and ensure the zip is created.

Change-Id: I8ca81771ab8740475626cb648674aefd3a841425