Automatically generate api diffs
We are now able to generate api diffs for a certain library.

Added a task that generates api diffs for any specific project
(generateLocalApiDiffs)

Also added a task that zips the generated diffs
(generateLocalApiDiffsArchive) therefore creating a ready to
upload api diffs archive.

Test: run the generateLocalDiffs task on a library and see
the diffs being created. Run the the generateApiDiffsArchive task
and check that the zip has been created.

Change-Id: I8ca81771ab8740475626cb648674aefd3a841425