All released apis are now recorded
UpdateApiTask now creates an api file for each release in a minor
release (alphas, betas and rcs) instead of having the same
final api file called version-major.version-minor.0.txt.
current.txt will still be updated all the time along with each
of the versioned txt file as per aosp/776005.

The reason for this is that we would like to preserve all api
changes for diff purposes as it would be nice to generate diffs for
x.y.z-alpha01->x.y.z-alpha02 and same for betas with forced
releases since we are currently unable to do that because we don't
create separate api files.

Comparing the api surface will only compare against the last final
api release (beta and above) as per aosp/746228.

Test: manual testing through the terminal ensuring checkApi and
updateApi work as expected.

Change-Id: I70aef482b8a470b63a956ca0bbb159b11275cdbb