Update patch set 2
Patch Set 2:

Only reason this is written in python is because it's intended to be expanded as part of my current work. The next addition would be to automate the upload to buganizer as I keep trying to find a way to include attachement. In short, this is intended to grow. Also I did not create the gradle task and make the docs depend on it precisely because I did not want every single local build to run it on each and every project (and also treehugger...) This is "seasonal", meaning a TL will run this when they are releasing and that is it, and if for some reason it breaks, fixing it will be simple and in the meantime it will not cause anyone else's build to fail. In short, it's expandable and it's seasonal and it won't cause any breaks and it won't make build times in slightly longer.

Patch-set: 2
Reviewer: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>